Improve accuracy of bounding box drawing
We take the final rendered dimensions of the expanded image
as a reference for where we draw the bounding box, instead of
assuming a fixed size. The difference will be more noticeable
on different dimension phones and different dimension images.

Also, clipped dimensions to [0, 1] before sending it to local model
as anything below 0 or above 1 does not make sense to train on.